+++ packages/terra-action-header/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 2.93.0 - (April 4, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 2.92.0 - (March 20, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 2.91.0 - (February 27, 2024)
 
 * Changed