+++ packages/terra-button-group/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 3.79.0 - (April 4, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 3.78.0 - (March 20, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 3.77.0 - (February 27, 2024)
 
 * Changed