+++ packages/terra-form-select/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 6.61.0 - (April 4, 2024)
+
+* Fixed
+  * Fixed exception thrown when down arrow is pressed on a form select with no options.
+
+## 6.60.0 - (March 27, 2024)
+
+* Fixed
+  * Fixed focus indicators for orion-fusion-theme.
+
 ## 6.59.0 - (March 15, 2024)
 
 * Fixed