+++ packages/terra-toolbar/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 1.47.0 - (April 4, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 1.46.0 - (March 20, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 1.45.0 - (February 27, 2024)
 
 * Changed